Task 4.2 (commit 2) :
- add TestUtils class
