--- carservicerest/src/main/resources/messages.properties
@@ -1,3 +1,0 @@
-errors.400.title=The request contains errors
-errors.404.title=Nothing was found in the catalog
-errors.404.header=Error 404: Nothing was found
