--- carservicerest/src/main/resources/templates/cars.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"
-	xmlns:sec="http://www.thymeleaf.org/extras/spring-security"
-	xmlns:th="http://www.thymeleaf.org">
-<head>
-	<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-	<title>Courses Page</title>
-</head>
-<body style="max-width: 1000px; margin: 0 auto">
-	
-	<h2>Cars</h2>
-	
-	<table>
-		<thead>
-			<tr>
-				<th>Car ID</th>
-				<th>Car Code</th>
-				<th>Car Make</th>
-				<th>Car Model</th>
-				<th>Car Year</th>
-			</tr>
-		</thead>
-		<tbody>
-			<tr th:each="car : ${cars}">
-				<td th:text="${car.id}"></td>
-				<td th:text="${car.carCode}"></td>
-				<td th:text="${car.make}"></td>
-				<td th:text="${car.model}"></td>
-				<td th:text="${car.year}"></td>
-				<td>
-					<table>
-						<tbody>
-							<tr th:each="category : ${car.categories}">
-								<td th:text="${category.id}"></td>
-							</tr>
-						</tbody>
-					</table>
-				</td>
-			</tr>
-		</tbody>
-	</table>
-	
-</body>
-</html>
